Refactored packages, small updates to design
