+++ app/src/main/AndroidManifest.xml
@@ -10,7 +10,7 @@
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
 
     <application
-        android:name=".HackillinoisApp"
+        android:name=".App"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -26,9 +26,9 @@
             </intent-filter>
         </activity>
         <activity android:name=".ui.MainActivity" />
-        <activity android:name=".ui.login.GitHubLoginActivity" />
-        <activity android:name=".ui.login.LoginChooserActivity" />
-        <activity android:name=".ui.login.LoginActivity" />
+        <activity android:name=".ui.modules.login.GitHubLoginActivity" />
+        <activity android:name=".ui.modules.login.LoginChooserActivity" />
+        <activity android:name=".ui.modules.login.LoginActivity" />
     </application>
 
 </manifest>
