+++ app/src/main/java/org/hackillinois/android/helper/Utils.java
@@ -1,4 +1,4 @@
-package org.hackillinois.android;
+package org.hackillinois.android.helper;
 
 import android.content.Context;
 import android.net.ConnectivityManager;
