+++ app/src/main/java/org/hackillinois/android/ui/base/BaseActivity.java
@@ -1,4 +1,4 @@
-package org.hackillinois.android.ui;
+package org.hackillinois.android.ui.base;
 
 import android.content.Context;
 import android.support.v7.app.AppCompatActivity;
@@ -6,7 +6,7 @@ import android.support.v7.app.AppCompatActivity;
 import uk.co.chrisjenx.calligraphy.CalligraphyContextWrapper;
 
 
-public class HackillinoisActivity extends AppCompatActivity {
+public class BaseActivity extends AppCompatActivity {
 	@Override
 	protected void attachBaseContext(Context newBase) {
 		super.attachBaseContext(CalligraphyContextWrapper.wrap(newBase));
