+++ app/src/main/java/org/hackillinois/android/ui/base/BaseFragment.java
@@ -1,0 +1,6 @@
+package org.hackillinois.android.ui.base;
+
+import android.app.Fragment;
+
+public class BaseFragment extends Fragment {
+}
