--- app/src/main/java/org/hackillinois/android/ui/fragment/NotificationFragment.java
@@ -1,7 +1,0 @@
-package org.hackillinois.android.ui.fragment;
-
-import android.support.v4.app.Fragment;
-
-public class NotificationFragment extends Fragment {
-    
-}
