+++ app/src/main/java/org/hackillinois/android/ui/modules/notification/NotificationFragment.java
@@ -1,0 +1,7 @@
+package org.hackillinois.android.ui.modules.notification;
+
+import org.hackillinois.android.ui.base.BaseFragment;
+
+public class NotificationFragment extends BaseFragment {
+    
+}
