+++ app/src/main/java/org/hackillinois/android/ui/modules/profile/ProfileFragment.java
@@ -1,0 +1,7 @@
+package org.hackillinois.android.ui.modules.profile;
+
+import org.hackillinois.android.ui.base.BaseFragment;
+
+public class ProfileFragment extends BaseFragment {
+
+}
