+++ app/src/main/res/layout/activity_login.xml
@@ -9,7 +9,7 @@
     android:orientation="vertical"
     android:paddingEnd="41dp"
     android:paddingStart="41dp"
-    tools:context="org.hackillinois.android.ui.login.LoginActivity"
+    tools:context=".ui.modules.login.LoginActivity"
     tools:ignore="MissingPrefix">
 
     <LinearLayout
