+++ app/src/main/res/layout/dialog_event_info.xml
@@ -9,7 +9,7 @@
 
     <ImageButton
         android:id="@+id/event_info_close"
-        android:layout_width="wrap_content"
+        android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:layout_marginEnd="28dp"
         android:layout_marginTop="28dp"
@@ -20,31 +20,33 @@
 
     <ImageButton
         android:id="@+id/event_star"
-        android:layout_width="wrap_content"
+        android:layout_width="0dp"
         android:layout_height="wrap_content"
+        android:layout_marginBottom="8dp"
         android:layout_marginStart="38dp"
+        android:layout_marginTop="8dp"
         android:background="@android:color/transparent"
+        app:layout_constraintBottom_toBottomOf="@+id/event_summary_container"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/event_info_close"
+        app:layout_constraintTop_toTopOf="@+id/event_summary_container"
         app:srcCompat="@drawable/ic_star_border_lightblue_200_24dp"/>
 
     <LinearLayout
         android:id="@+id/event_summary_container"
-        android:layout_width="wrap_content"
+        android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:layout_marginEnd="8dp"
+        android:layout_marginStart="21dp"
         android:orientation="vertical"
         app:layout_constraintEnd_toStartOf="@+id/event_info_close"
-        app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintStart_toEndOf="@id/event_star"
-        app:layout_constraintTop_toTopOf="@id/event_info_close">
+        app:layout_constraintTop_toBottomOf="@id/event_info_close">
 
         <TextView
             android:id="@+id/event_name"
             fontPath="fonts/Roboto_Regular.ttf"
-            android:layout_width="wrap_content"
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_marginStart="21dp"
             android:fontFamily="sans-serif"
             android:textColor="@color/darkPurple"
             android:textSize="20sp"
@@ -56,7 +58,6 @@
             fontPath="fonts/Roboto_Bold.ttf"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginStart="21dp"
             android:fontFamily="sans-serif"
             android:textColor="@color/darkPurple"
             android:textSize="16sp"
@@ -68,7 +69,6 @@
             fontPath="fonts/Roboto_Bold.ttf"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginStart="21dp"
             android:fontFamily="sans-serif"
             android:textColor="@color/darkPurple"
             android:textSize="14sp"
@@ -78,12 +78,12 @@
 
     <TextView
         android:id="@+id/event_description"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
+        android:layout_width="0dp"
+        android:layout_height="0dp"
         android:layout_marginBottom="8dp"
-        android:layout_marginEnd="24dp"
-        android:layout_marginStart="38dp"
-        android:layout_marginTop="37dp"
+        android:layout_marginEnd="28dp"
+        android:layout_marginStart="28dp"
+        android:layout_marginTop="28dp"
         android:fontFamily="sans-serif"
         android:textColor="@color/darkPurple"
         android:textSize="13sp"
@@ -93,22 +93,23 @@
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/event_summary_container"
-        app:layout_constraintVertical_bias="0.0"
         tools:text="This event is the best event, like the best event, in the whole world. It’s the event of events and all events."/>
 
     <Button
         android:id="@+id/login_next"
         fontPath="fonts/Roboto_Medium.ttf"
-        android:layout_width="match_parent"
+        android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:layout_marginBottom="27dp"
-        android:layout_marginEnd="20dp"
-        android:layout_marginStart="20dp"
+        android:layout_marginBottom="28dp"
+        android:layout_marginEnd="28dp"
+        android:layout_marginStart="28dp"
         android:background="@color/bluePurple"
         android:fontFamily="sans-serif-medium"
-        android:text="@string/next"
+        android:text="@string/get_directions"
         android:textColor="@color/darkPurple"
         android:textSize="14sp"
         android:textStyle="normal"
-        app:layout_constraintBottom_toBottomOf="parent"/>
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"/>
 </android.support.constraint.ConstraintLayout>
