+++ app/src/main/res/values/strings.xml
@@ -29,4 +29,5 @@
     <string name="sunday">Sunday</string>
     <string name="saturday">Saturday</string>
     <string name="friday">Friday</string>
+    <string name="get_directions">Get Directions</string>
 </resources>
